Fixed most of the pmd errors

+++ src/main/java/com/thelegendofbald/api/settingsmenu/SettingOption.java
@@ -39,7 +39,7 @@ public interface SettingOption {
      * @return the number of setting options as an integer
      */
     static int getSize() {
-        return (int) Settings.getMaxIndex();
+        return Settings.getMaxIndex();
     }
 
 }
